Remove `include/grpcpp/impl/codegen/server_callback.h` (#31280)
