+++ include/grpcpp/ext/server_load_reporting.h
@@ -22,8 +22,8 @@
 #include <grpc/support/port_platform.h>
 
 #include <grpc/load_reporting.h>
-#include <grpcpp/impl/codegen/server_context.h>
 #include <grpcpp/impl/server_builder_option.h>
+#include <grpcpp/server_context.h>
 #include <grpcpp/support/config.h>
 
 namespace grpc {
