+++ src/compiler/cpp_generator.cc
@@ -149,7 +149,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/impl/rpc_method.h",
         "grpcpp/impl/codegen/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
-        "grpcpp/impl/codegen/server_context.h",
+        "grpcpp/server_context.h",
         "grpcpp/impl/codegen/service_type.h",
         "grpcpp/impl/codegen/status.h",
         "grpcpp/impl/codegen/stub_options.h",
@@ -1658,7 +1658,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/impl/codegen/rpc_service_method.h",
         "grpcpp/impl/codegen/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
-        "grpcpp/impl/codegen/server_context.h",
+        "grpcpp/server_context.h",
         "grpcpp/impl/codegen/service_type.h",
         "grpcpp/impl/codegen/sync_stream.h"};
     std::vector<std::string> headers(headers_strs, array_end(headers_strs));
